# artifacts/api-server/dist/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: VetStan Veterinary Dictionary Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /diseases:
    get:
      operationId: listDiseases
      tags: [diseases]
      summary: List all diseases
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/SearchParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedDiseases"
    post:
      operationId: createDisease
      tags: [diseases]
      summary: Create a disease
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDiseaseBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Disease"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /diseases/{id}:
    get:
      operationId: getDisease
      tags: [diseases]
      summary: Get a disease by ID
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Disease"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateDisease
      tags: [diseases]
      summary: Update a disease
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDiseaseBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Disease"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteDisease
      tags: [diseases]
      summary: Delete a disease
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # --- DRUGS ---
...
  /instruments:
    get:
      operationId: listInstruments
      tags: [instruments]
      summary: List all instruments
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/SearchParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedInstruments"
    post:
      operationId: createInstrument
      tags: [instruments]
      summary: Create an instrument
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInstrumentBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Instrument"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /instruments/{id}:
    get:
      operationId: getInstrument
      tags: [instruments]
      summary: Get an instrument by ID
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Instrument"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateInstrument
      tags: [instruments]
      summary: Update an instrument
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInstrumentBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Instrument"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteInstrument
      tags: [instruments]
      summary: Delete an instrument
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # --- TESTS ---
...
  /normal-ranges:
    get:
      operationId: listNormalRanges
      tags: [normalRanges]
      summary: List all normal ranges
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/SearchParam"
        - name: species
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedNormalRanges"
    post:
      operationId: createNormalRange
      tags: [normalRanges]
      summary: Create a normal range
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNormalRangeBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NormalRange"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /normal-ranges/{id}:
    get:
      operationId: getNormalRange
      tags: [normalRanges]
      summary: Get a normal range by ID
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NormalRange"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateNormalRange
      tags: [normalRanges]
      summary: Update a normal range
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNormalRangeBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NormalRange"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteNormalRange
      tags: [normalRanges]
      summary: Delete a normal range
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # --- ABOUT ---
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
    SearchParam:
      name: q
      in: query
      schema:
        type: string
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        pages:
          type: integer
      required:
        - page
        - limit
        - total
        - pages

    # --- AUTH ---
...
    Disease:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        kurdish:
          type: ["string", "null"]
        symptoms:
          type: ["string", "null"]
        cause:
          type: ["string", "null"]
        control:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - createdAt
        - updatedAt
    CreateDiseaseBody:
      type: object
      properties:
        name:
          type: string
        kurdish:
          type: string
        symptoms:
          type: string
        cause:
          type: string
        control:
          type: string
      required:
        - name
    PaginatedDiseases:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Disease"
        pagination:
          $ref: "#/components/schemas/Pagination"
      required:
        - items
        - pagination

    # --- DRUG ---
...
    Instrument:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        usage:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        manufacturer:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - createdAt
        - updatedAt
    CreateInstrumentBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        usage:
          type: string
        category:
          type: string
        manufacturer:
          type: string
      required:
        - name
    PaginatedInstruments:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Instrument"
        pagination:
          $ref: "#/components/schemas/Pagination"
      required:
        - items
        - pagination

    # --- TEST ---
...
    NormalRange:
      type: object
      properties:
        id:
          type: integer
        parameter:
          type: string
        species:
          type: string
        category:
          type: ["string", "null"]
        minValue:
          type: ["number", "null"]
        maxValue:
          type: ["number", "null"]
        unit:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - parameter
        - species
        - createdAt
        - updatedAt
    CreateNormalRangeBody:
      type: object
      properties:
        parameter:
          type: string
        species:
          type: string
        category:
          type: string
        minValue:
          type: number
        maxValue:
          type: number
        unit:
          type: string
        description:
          type: string
      required:
        - parameter
        - species
    PaginatedNormalRanges:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/NormalRange"
        pagination:
          $ref: "#/components/schemas/Pagination"
      required:
        - items
        - pagination

    # --- CEO ---
...